refactor: impl EventWrapper
